Introduce block_synchronizer_handler and wire to block waiter (part 1) #214 (#227)
feat: wire block_synchronizer to block_waiter

This commit is introducing the block synchronizer handler which is a sync facade for the block synchronizer. We are using the handler on the block_waiter to ensure that certificates are synced during when we retrieve blocks. Also, for certificates fetched on the fly from peers, they will be submitted for further processing & validation on core to ensure amongst others parent causal completion.